requirements.txt
# Install PyTorch first following https://pytorch.org/get-started/locally/
# Example (CPU only):
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
# Example (CUDA 12.1):
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

datasets>=2.18.0
pillow>=10.0.0
requests>=2.31.0
tqdm>=4.66.0
scikit-learn>=1.3.0
pandas>=2.1.0
matplotlib>=3.8.0
pyyaml>=6.0.1
streamlit>=1.35.0
